+++ packages/types/src/elements/datePicker/index.ts
@@ -3,7 +3,8 @@ import {
   GenericLezzformElement,
   LezzformElementCategory,
 } from "../..";
-import { DateRule, GeneralRules } from "../../rules";
+import { DateRule } from "../../rules";
+import { ElementAttributesType } from "../../shared";
 import { GeneralStyleMetadata } from "../../styles";
 
 export type DatePickerType = GenericLezzformElement<
@@ -15,10 +16,24 @@ export type DatePickerType = GenericLezzformElement<
   rule: DatePickerRuleType;
 };
 
-export type DatePickerAttributesType = {
+export type DatePickerAttributesType = ElementAttributesType & {
   format: string;
 };
 
-export type DatePickerStyleType = GeneralStyleMetadata;
+export type DatePickerStyleType = GeneralStyleMetadata &
+  DatePickerStyleMetadataValuesType;
+
+export type DatePickerStyleMetadataValuesType = {
+  input: Record<string, unknown>;
+  prefixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+  suffixAdornment: {
+    icon?: Record<string, unknown>;
+    text?: Record<string, unknown>;
+  };
+  content: Record<string, unknown>;
+};
 
 export type DatePickerRuleType = DateRule;